用户登录界面完成

+++ src/main/java/com/chengxiaoxiao/seckillshop/domain/User.java
@@ -1,41 +1,30 @@
 package com.chengxiaoxiao.seckillshop.domain;
 
-/**
- * @author XiaoXiao
- * @version $Rev$
- */
-public class User
-{
-    private long id;
+public class User {
+    private int id;
     private String name;
 
-    public User()
-    {
+    public User() {
     }
 
-    public User(long id, String name)
-    {
+    public User(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId()
-    {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id)
-    {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public String getName()
-    {
+    public String getName() {
         return name;
     }
 
-    public void setName(String name)
-    {
+    public void setName(String name) {
         this.name = name;
     }
 }
